morre

+++ My project/Assets/Scripts/IDamageable.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    void TakeDamage(int damageAmount);
+    void TakeDamage(float damageAmount);
 }
